Report runtime test coverage
Drafting as I'm still working on where to show the reports. You can run locally by doing:
```
cd pacakges/runtime/tests/unit_tests
./test.sh coverage
```

Diffs=
99cf88f77d Report runtime test coverage (#9114)

+++ src/animation/animation_reset_factory.cpp
@@ -114,7 +114,8 @@ public:
             if (propertiesData.size() > 0)
             {
                 animationReset->writeObjectId(keyedObjectData->objectId);
-                animationReset->writeTotalProperties(propertiesData.size());
+                animationReset->writeTotalProperties(
+                    (uint32_t)propertiesData.size());
                 for (auto keyedPropertyData : propertiesData)
                 {
                     auto keyedProperty = keyedPropertyData.keyedProperty;
